merge dev

+++ SolastaUnfinishedBusiness/DataViewer/ReflectionTreeView.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using SolastaUnfinishedBusiness.Api.Infrastructure;
+using SolastaUnfinishedBusiness.Api.LanguageExtensions;
 using SolastaUnfinishedBusiness.Api.ModKit;
 using SolastaUnfinishedBusiness.Displays;
 using UnityEngine;
@@ -141,7 +141,7 @@ internal class ReflectionTreeView
                 {
                     searchText = searchText.Trim();
                     ReflectionSearch.Shared.StartSearch(_tree.RootNode, searchText, UpdateCounts, _searchResults);
-                }, UI.Width(250));
+                }, UI.Width((float)250));
                 GUILayout.Space(10f);
                 var isSearching = ReflectionSearch.Shared.IsSearching;
                 UI.ActionButton(isSearching ? "Stop" : "Search", () =>
@@ -275,7 +275,8 @@ internal class ReflectionTreeView
                 name + " : " + (
                     node.IsBaseType ? node.Type.Name.Grey() :
                     node.IsGameObject ? node.Type.Name.Magenta() :
-                    node.IsEnumerable ? node.Type.Name.Cyan() : node.Type.Name.Orange()), GUILayout.ExpandWidth(false),
+                    node.IsEnumerable ? node.Type.Name.Cyan() : node.Type.Name.Orange()),
+                GUILayout.ExpandWidth(false),
                 GUILayout.MinWidth(TitleMinWidth));
 
             // value
