+++ SolastaUnfinishedBusiness/Displays/PatchesDisplay.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Reflection;
 using HarmonyLib;
 using JetBrains.Annotations;
-using SolastaUnfinishedBusiness.Api.Infrastructure;
+using SolastaUnfinishedBusiness.Api.LanguageExtensions;
 using SolastaUnfinishedBusiness.Api.ModKit;
 using UnityEngine;
 using Random = UnityEngine.Random;
@@ -56,17 +56,17 @@ internal static class PatchesDisplay
                     using (new GUILayout.HorizontalScope())
                     {
                         GUILayout.Label($"{index++}", GUI.skin.box, UI.AutoWidth());
-                        UI.Space(10);
+                        UI.Space((float)10);
                         GUILayout.Label(
                             $"{returnTypeStr.Grey().Bold()} {methodName.Bold()}\t{typeStr.Grey().Italic()}");
                     }
 
                     var patches = EnabledPatchesForMethod(method);
 
-                    UI.Space(15);
+                    UI.Space((float)15);
                     using (new GUILayout.HorizontalScope())
                     {
-                        UI.Space(50);
+                        UI.Space((float)50);
 
                         using (new GUILayout.VerticalScope())
                         {
@@ -76,17 +76,18 @@ internal static class PatchesDisplay
                             }
                         }
 
-                        UI.Space(10);
+                        UI.Space((float)10);
                         using (new GUILayout.VerticalScope())
                         {
                             foreach (var patch in patches)
                             {
-                                GUILayout.Label(patch.owner.Color($"#{_modIdsToColor[patch.owner]}").Bold(),
+                                GUILayout.Label(
+                                    patch.owner.Color($"#{_modIdsToColor[patch.owner]}").Bold(),
                                     GUI.skin.label);
                             }
                         }
 
-                        UI.Space(10);
+                        UI.Space((float)10);
                         using (new GUILayout.VerticalScope())
                         {
                             foreach (var patch in patches)
@@ -95,7 +96,7 @@ internal static class PatchesDisplay
                             }
                         }
 
-                        UI.Space(10);
+                        UI.Space((float)10);
                         using (new GUILayout.VerticalScope())
                         {
                             foreach (var patch in patches)
@@ -108,7 +109,7 @@ internal static class PatchesDisplay
                             }
                         }
 
-                        UI.Space(10);
+                        UI.Space((float)10);
                         using (new GUILayout.VerticalScope())
                         {
                             foreach (var patch in patches)
